refactor: address local HTTPS feedback
