+++ .changeset/local-https-vite-plugin.md
@@ -2,4 +2,4 @@
 "@shopify/hydrogen": minor
 ---
 
-Add `localHttps()` and `localHttpsDevServer()` under `@shopify/hydrogen/vite` for portable local HTTPS development with Customer Account API flows.
+Add `localHttps()` under `@shopify/hydrogen/vite` for portable local HTTPS development with Customer Account API flows. Frameworks that terminate HTTPS outside Vite can use `localHttps(...).api.getDevServerConfig()`.
