+++ examples/astro/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "astro dev",
-    "https:dev": "VITE_LOCAL_HTTPS=1 astro dev",
+    "https:dev": "astro dev",
     "build": "astro build",
     "start": "node ./dist/server/entry.mjs",
     "typecheck": "astro check"
