+++ examples/nuxt/nuxt.config.ts
@@ -1,15 +1,17 @@
-import { localHttps, localHttpsDevServer } from "@shopify/hydrogen/vite";
+import { localHttps } from "@shopify/hydrogen/vite";
 import tailwindcss from "@tailwindcss/vite";
 import type { NuxtConfig } from "nuxt/schema";
 
 type VitePlugin = NonNullable<NonNullable<NuxtConfig["vite"]>["plugins"]>[number];
 
-const enabled = process.env.VITE_LOCAL_HTTPS === "1";
+const enabled =
+  process.env.VITE_LOCAL_HTTPS === "1" || process.env.npm_lifecycle_event === "https:dev";
 const httpsOptions = { enabled };
+const httpsPlugin = localHttps(httpsOptions);
 
 export default defineNuxtConfig({
   compatibilityDate: "2025-05-08",
-  devServer: localHttpsDevServer(httpsOptions),
+  devServer: httpsPlugin.api.getDevServerConfig(),
   alias: {
     "@shared": new URL("../shared", import.meta.url).pathname,
   },
@@ -23,6 +25,6 @@ export default defineNuxtConfig({
   css: ["~/assets/css/main.css"],
   ssr: true,
   vite: {
-    plugins: [localHttps(httpsOptions) as VitePlugin, tailwindcss() as VitePlugin],
+    plugins: [httpsPlugin as VitePlugin, tailwindcss() as VitePlugin],
   },
 });
