+++ examples/nuxt/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "nuxt dev",
-    "https:dev": "VITE_LOCAL_HTTPS=1 nuxt dev",
+    "https:dev": "nuxt dev",
     "build": "nuxt build",
     "start": "node .output/server/index.mjs",
     "typecheck": "nuxt typecheck"
