+++ examples/solid-start/app.config.ts
@@ -1,10 +1,12 @@
-import { localHttps, localHttpsDevServer } from "@shopify/hydrogen/vite";
+import { localHttps } from "@shopify/hydrogen/vite";
 import { defineConfig } from "@solidjs/start/config";
 import tailwindcss from "@tailwindcss/vite";
 
-const enabled = process.env.VITE_LOCAL_HTTPS === "1";
+const enabled =
+  process.env.VITE_LOCAL_HTTPS === "1" || process.env.npm_lifecycle_event === "https:dev";
 const httpsOptions = { enabled };
-const devServer = localHttpsDevServer(httpsOptions);
+const httpsPlugin = localHttps(httpsOptions);
+const devServer = httpsPlugin.api.getDevServerConfig();
 
 export default defineConfig({
   middleware: "src/middleware.ts",
@@ -12,7 +14,7 @@ export default defineConfig({
     https: devServer?.https,
   },
   vite: {
-    plugins: [localHttps(httpsOptions), tailwindcss()],
+    plugins: [httpsPlugin, tailwindcss()],
     resolve: {
       alias: {
         "@shared": new URL("../shared", import.meta.url).pathname,
