+++ packages/hydrogen/src/customer-account/paths.ts
@@ -1,0 +1,4 @@
+export const CUSTOMER_ACCOUNT_AUTHORIZE_PATH = "/account/authorize" as const;
+export const CUSTOMER_ACCOUNT_LOGIN_PATH = "/account/login" as const;
+export const CUSTOMER_ACCOUNT_LOGOUT_PATH = "/account/logout" as const;
+export const CUSTOMER_ACCOUNT_REFRESH_PATH = "/account/refresh" as const;
