+++ packages/hydrogen/src/customer-account/session.ts
@@ -8,11 +8,19 @@ import {
   type ShopifyRouteSessionManager,
 } from "../core/request-routing/registered-routes";
 import { CustomerAccountApiError, CustomerAccountOAuthError } from "./errors";
+import {
+  CUSTOMER_ACCOUNT_AUTHORIZE_PATH,
+  CUSTOMER_ACCOUNT_LOGIN_PATH,
+  CUSTOMER_ACCOUNT_LOGOUT_PATH,
+  CUSTOMER_ACCOUNT_REFRESH_PATH,
+} from "./paths";
 
-export const CUSTOMER_ACCOUNT_AUTHORIZE_PATH = "/account/authorize" as const;
-export const CUSTOMER_ACCOUNT_LOGIN_PATH = "/account/login" as const;
-export const CUSTOMER_ACCOUNT_LOGOUT_PATH = "/account/logout" as const;
-export const CUSTOMER_ACCOUNT_REFRESH_PATH = "/account/refresh" as const;
+export {
+  CUSTOMER_ACCOUNT_AUTHORIZE_PATH,
+  CUSTOMER_ACCOUNT_LOGIN_PATH,
+  CUSTOMER_ACCOUNT_LOGOUT_PATH,
+  CUSTOMER_ACCOUNT_REFRESH_PATH,
+} from "./paths";
 
 const CUSTOMER_ACCOUNT_SESSION_KEY = "customerAccount";
 const DEFAULT_LOGIN_RETURN_TO_PATH = "/account";
