+++ templates/react-router/package.json
@@ -7,7 +7,7 @@
     "build": "react-router build",
     "deploy": "shopify hydrogen deploy --assets-dir dist/client --worker-dir dist/server",
     "dev": "vite dev",
-    "https:dev": "VITE_LOCAL_HTTPS=1 vite dev",
+    "https:dev": "vite dev",
     "preview": "react-router build && vite preview",
     "typecheck": "react-router typegen && tsc && hydrogen gql check --fail-on-warn"
   },
