+++ templates/react-router/vite.config.ts
@@ -6,7 +6,8 @@ import { defineConfig } from "vite";
 
 const oxygenPlugins = oxygen();
 const oxygenPlugin = oxygenPlugins.find((plugin) => plugin.name === "oxygen:main");
-const enabled = process.env.VITE_LOCAL_HTTPS === "1";
+const enabled =
+  process.env.VITE_LOCAL_HTTPS === "1" || process.env.npm_lifecycle_event === "https:dev";
 const httpsOptions = { enabled };
 
 if (!oxygenPlugin?.api) {
